+++ changes.md
@@ -10,6 +10,7 @@
 - Adding _f144_, _al00_ and _ep01_ writer modules. For more information on the schemas mentioned,
   see ([schema definitions here](https://github.com/ess-dmsc/streaming-data-types)).
 - Adding _se00_ writer module (see [schema definitions here](https://github.com/ess-dmsc/streaming-data-types)).
+- Adding _ev44_ writer module (see [schema definitions here](https://github.com/ess-dmsc/streaming-data-types)).
 
 
 ## Version 5.2.0: Kafka improvements and other fixes